requirements.txt
requests
from amadeus import Client, ResponseError
python-dotenv